+++ __init__.py
@@ -2,20 +2,9 @@ from .py import nodes, powernoise, sonar
 
 sonar.add_samplers()
 
-NODE_CLASS_MAPPINGS = {
-    "SamplerSonarEuler": nodes.SamplerNodeSonarEuler,
-    "SamplerSonarEulerA": nodes.SamplerNodeSonarEulerAncestral,
-    "SamplerSonarDPMPPSDE": nodes.SamplerNodeSonarDPMPPSDE,
-    "SamplerConfigOverride": nodes.SamplerNodeConfigOverride,
-    "NoisyLatentLike": nodes.NoisyLatentLikeNode,
-    "SonarCustomNoise": nodes.SonarCustomNoiseNode,
+NODE_CLASS_MAPPINGS = nodes.NODE_CLASS_MAPPINGS | {
     "SonarPowerNoise": powernoise.SonarPowerNoiseNode,
-    "SonarGuidanceConfig": nodes.GuidanceConfigNode,
 }
-
-NODE_DISPLAY_NAME_MAPPINGS = {}
-
-if hasattr(nodes, "KRestartSamplerCustomNoise"):
-    NODE_CLASS_MAPPINGS["KRestartSamplerCustomNoise"] = nodes.KRestartSamplerCustomNoise
+NODE_DISPLAY_NAME_MAPPINGS = nodes.NODE_DISPLAY_NAME_MAPPINGS
 
 __all__ = ["NODE_CLASS_MAPPINGS", "NODE_DISPLAY_NAME_MAPPINGS"]